+++ packages/xchain-ethereum/src/client.ts
@@ -14,6 +14,10 @@ import {
   FeesWithGasPricesAndLimits,
   InfuraCreds,
   ApproveParams,
+  CallParams,
+  IsApprovedParams,
+  EstimateCallParams,
+  EstimateApproveParams,
 } from './types'
 import {
   Address,
@@ -31,7 +35,6 @@ import {
   FeeOptionKey,
   FeesParams as XFeesParams,
   BaseXChainClient,
-  FeeRates,
 } from '@xchainjs/xchain-client'
 import { AssetETH, baseAmount, BaseAmount, assetToString, Asset, delay } from '@xchainjs/xchain-util'
 import * as ethplorerAPI from './ethplorer-api'
@@ -57,30 +60,14 @@ import {
  */
 export interface EthereumClient {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  call<T>(
-    walletIndex: number,
-    asset: Address,
-    abi: ethers.ContractInterface,
-    func: string,
-    params: Array<unknown>,
-  ): Promise<T>
-  estimateCall(
-    asset: Address,
-    abi: ethers.ContractInterface,
-    func: string,
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    params: Array<any>,
-  ): Promise<BigNumber>
+  call<T>(params: CallParams): Promise<T>
+  estimateCall(asset: EstimateCallParams): Promise<BigNumber>
   estimateGasPrices(): Promise<GasPrices>
   estimateGasLimit(params: FeesParams): Promise<BigNumber>
   estimateFeesWithGasPricesAndLimits(params: FeesParams): Promise<FeesWithGasPricesAndLimits>
-
-  isApproved(spender: Address, sender: Address, amount: BaseAmount): Promise<boolean>
-  approve(
-    params: ApproveParams & {
-      feeOptionKey?: FeeOptionKey
-    },
-  ): Promise<TransactionResponse>
+  estimateApprove(params: EstimateApproveParams): Promise<BigNumber>
+  isApproved(params: IsApprovedParams): Promise<boolean>
+  approve(params: ApproveParams): Promise<TransactionResponse>
 }
 
 export type EthereumClientParams = XChainClientParams & {
@@ -156,28 +143,30 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
   /**
    * Get the current address.
    *
+   * @param {number} walletIndex (optional) HD wallet index
    * @returns {Address} The current address.
    *
    * @throws {"Phrase must be provided"}
    * Thrown if phrase has not been set before. A phrase is needed to create a wallet and to derive an address from it.
    */
-  getAddress(index = 0): Address {
-    if (index < 0) {
+  getAddress(walletIndex = 0): Address {
+    if (walletIndex < 0) {
       throw new Error('index must be greater than zero')
     }
-    return this.hdNode.derivePath(this.getFullDerivationPath(index)).address.toLowerCase()
+    return this.hdNode.derivePath(this.getFullDerivationPath(walletIndex)).address.toLowerCase()
   }
 
   /**
    * Get etherjs wallet interface.
    *
+   * @param {number} walletIndex (optional) HD wallet index
    * @returns {Wallet} The current etherjs wallet interface.
    *
    * @throws {"Phrase must be provided"}
    * Thrown if phrase has not been set before. A phrase is needed to create a wallet and to derive an address from it.
    */
-  getWallet = (index = 0): ethers.Wallet => {
-    return new Wallet(this.hdNode.derivePath(this.getFullDerivationPath(index))).connect(this.getProvider())
+  getWallet = (walletIndex = 0): ethers.Wallet => {
+    return new Wallet(this.hdNode.derivePath(this.getFullDerivationPath(walletIndex))).connect(this.getProvider())
   }
   setupProviders = (): void => {
     if (this.infuraCreds) {
@@ -284,6 +273,7 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
    * Set/update a new phrase (Eg. If user wants to change wallet)
    *
    * @param {string} phrase A new phrase.
+   * @param {number} walletIndex (optional) HD wallet index
    * @returns {Address} The address from the given phrase
    *
    * @throws {"Invalid phrase"}
@@ -359,8 +349,9 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
               apiKey: etherscan.apiKey,
             })
             const decimals =
-              BigNumber.from(await this.call<BigNumberish>(0, assetAddress, erc20ABI, 'decimals', [])).toNumber() ||
-              ETH_DECIMAL
+              BigNumber.from(
+                await this.call<BigNumberish>({ contractAddress: assetAddress, abi: erc20ABI, funcName: 'decimals' }),
+              ).toNumber() || ETH_DECIMAL
 
             if (!Number.isNaN(decimals)) {
               balances.push({
@@ -506,137 +497,147 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
   /**
    * Call a contract function.
    * @template T The result interface.
-   * @param {Address} address The contract address.
+   * @param {number} walletIndex (optional) HD wallet index
+   * @param {Address} contractAddress The contract address.
    * @param {ContractInterface} abi The contract ABI json.
-   * @param {string} func The function to be called.
-   * @param {Array<any>} params The parameters of the function.
+   * @param {string} funcName The function to be called.
+   * @param {Array<any>} funcParams The parameters of the function.
    * @returns {T} The result of the contract function call.
    *
-   * @throws {"address must be provided"}
+   * @throws {"contracAddress must be provided"}
    * Thrown if the given contract address is empty.
    */
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  call = async <T>(
-    walletIndex = 0,
-    contractAddress: Address,
-    abi: ethers.ContractInterface,
-    func: string,
-    params: Array<unknown>,
-  ): Promise<T> => {
+  call = async <T>({ walletIndex = 0, contractAddress, abi, funcName, funcParams = [] }: CallParams): Promise<T> => {
     if (!contractAddress) {
       return Promise.reject(new Error('contractAddress must be provided'))
     }
     const contract = new ethers.Contract(contractAddress, abi, this.getProvider()).connect(this.getWallet(walletIndex))
-    return contract[func](...params)
+    return contract[funcName](...funcParams)
   }
 
   /**
    * Call a contract function.
-   * @param {Address} address The contract address.
+   * @param {Address} contractAddress The contract address.
    * @param {ContractInterface} abi The contract ABI json.
-   * @param {string} func The function to be called.
-   * @param {Array<any>} params The parameters of the function.
+   * @param {string} funcName The function to be called.
+   * @param {Array<any>} funcParams The parameters of the function.
+   * @param {number} walletIndex (optional) HD wallet index
    * @returns {BigNumber} The result of the contract function call.
    *
    * @throws {"address must be provided"}
    * Thrown if the given contract address is empty.
    */
-  estimateCall = async (
-    contractAddress: Address,
-    abi: ethers.ContractInterface,
-    func: string,
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    params: Array<any>,
-  ): Promise<BigNumber> => {
+  estimateCall = async ({
+    contractAddress,
+    abi,
+    funcName,
+    funcParams = [],
+    walletIndex = 0,
+  }: EstimateCallParams): Promise<BigNumber> => {
     if (!contractAddress) {
       return Promise.reject(new Error('contractAddress must be provided'))
     }
-    const contract = new ethers.Contract(contractAddress, abi, this.getProvider()).connect(this.getWallet(0))
-    return contract.estimateGas[func](...params)
+    const contract = new ethers.Contract(contractAddress, abi, this.getProvider()).connect(this.getWallet(walletIndex))
+    return contract.estimateGas[funcName](...funcParams)
   }
 
   /**
    * Check allowance.
    *
-   * @param {Address} spender The spender address.
-   * @param {Address} sender The sender address.
-   * @param {BaseAmount} amount The amount of token.
+   * @param {Address} contractAddress The spender address.
+   * @param {Address} spenderAddress The spender address.
+   * @param {BaseAmount} amount The amount to check if it's allowed to spend or not (optional).
+   * @param {number} walletIndex (optional) HD wallet index
    * @returns {boolean} `true` or `false`.
    */
-  isApproved = async (spender: Address, sender: Address, amount: BaseAmount): Promise<boolean> => {
-    try {
-      const txAmount = BigNumber.from(amount.amount().toFixed())
-      const allowance = await this.call<BigNumberish>(0, sender, erc20ABI, 'allowance', [spender, spender])
-      return txAmount.lte(allowance)
-    } catch (error) {
-      return Promise.reject(error)
-    }
+  isApproved = async ({
+    contractAddress,
+    spenderAddress,
+    amount,
+    walletIndex = 0,
+  }: IsApprovedParams): Promise<boolean> => {
+    // since amount is optional, set it to smallest amount by default
+    const txAmount = BigNumber.from(amount?.amount().toFixed() ?? 1)
+    const owner = this.getAddress(walletIndex)
+    const allowance = await this.call<BigNumberish>({
+      contractAddress,
+      abi: erc20ABI,
+      funcName: 'allowance',
+      funcParams: [owner, spenderAddress],
+    })
+    return txAmount.lte(allowance)
   }
 
   /**
    * Check allowance.
    *
-   * @param {number} walletIndex which wallet to use to make the call
-   * @param {Address} spender The spender index.
-   * @param {Address} sender The sender address.
+   * @param {Address} contractAddress The contract address.
+   * @param {Address} spenderAddress The spender address.
    * @param {feeOptionKey} FeeOptionKey Fee option (optional)
    * @param {BaseAmount} amount The amount of token. By default, it will be unlimited token allowance. (optional)
+   * @param {number} walletIndex (optional) HD wallet index
+   *
    * @returns {TransactionResponse} The transaction result.
    */
   approve = async ({
-    walletIndex = 0,
-    spender,
-    sender,
-    feeOptionKey,
+    contractAddress,
+    spenderAddress,
+    feeOptionKey = 'fastest',
     amount,
+    walletIndex = 0,
+    gasLimitFallback,
   }: ApproveParams): Promise<TransactionResponse> => {
-    const gasPrice =
-      feeOptionKey &&
-      BigNumber.from(
-        (
-          await this.estimateGasPrices()
-            .then((prices) => prices[feeOptionKey])
-            .catch(() => getDefaultGasPrices()[feeOptionKey])
-        )
-          .amount()
-          .toFixed(),
+    const gasPrice = BigNumber.from(
+      (
+        await this.estimateGasPrices()
+          .then((prices) => prices[feeOptionKey])
+          .catch(() => getDefaultGasPrices()[feeOptionKey])
       )
-    const gasLimit = await this.estimateApprove({ spender, sender, amount }).catch(() => undefined)
+        .amount()
+        .toFixed(),
+    )
+    const gasLimit = await this.estimateApprove({
+      walletIndex,
+      spenderAddress,
+      contractAddress,
+      amount,
+    }).catch(() => BigNumber.from(gasLimitFallback))
 
-    try {
-      const txAmount = amount ? BigNumber.from(amount.amount().toFixed()) : MAX_APPROVAL
-      const txResult = await this.call<TransactionResponse>(walletIndex, sender, erc20ABI, 'approve', [
-        spender,
-        txAmount,
-        { from: this.getAddress(), gasPrice, gasLimit },
-      ])
-
-      return txResult
-    } catch (error) {
-      return Promise.reject(error)
-    }
+    const txAmount = amount ? BigNumber.from(amount.amount().toFixed()) : MAX_APPROVAL
+    return await this.call<TransactionResponse>({
+      walletIndex,
+      contractAddress,
+      abi: erc20ABI,
+      funcName: 'approve',
+      funcParams: [spenderAddress, txAmount, { from: this.getAddress(walletIndex), gasPrice, gasLimit }],
+    })
   }
 
   /**
    * Estimate gas limit of approve.
    *
-   * @param {Address} spender The spender address.
-   * @param {Address} sender The sender address.
+   * @param {Address} contractAddress The contract address.
+   * @param {Address} spenderAddress The spender address.
+   * @param {number} walletIndex (optional) HD wallet index
    * @param {BaseAmount} amount The amount of token. By default, it will be unlimited token allowance. (optional)
    * @returns {BigNumber} The estimated gas limit.
    */
   estimateApprove = async ({
-    spender,
-    sender,
+    contractAddress,
+    spenderAddress,
+    walletIndex = 0,
     amount,
-  }: Omit<ApproveParams, 'feeOptionKey' | 'walletIndex'>): Promise<BigNumber> => {
+  }: EstimateApproveParams): Promise<BigNumber> => {
     try {
       const txAmount = amount ? BigNumber.from(amount.amount().toFixed()) : MAX_APPROVAL
-      const gasLimit = await this.estimateCall(sender, erc20ABI, 'approve', [
-        spender,
-        txAmount,
-        { from: this.getAddress() },
-      ])
+      const gasLimit = await this.estimateCall({
+        walletIndex,
+        contractAddress,
+        abi: erc20ABI,
+        funcName: 'approve',
+        funcParams: [spenderAddress, txAmount, { from: this.getAddress(walletIndex) }],
+      })
 
       return gasLimit
     } catch (error) {
@@ -708,11 +709,13 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
       let txResult
       if (assetAddress && !isETHAddress) {
         // Transfer ERC20
-        txResult = await this.call<TransactionResponse>(walletIndex, assetAddress, erc20ABI, 'transfer', [
-          recipient,
-          txAmount,
-          Object.assign({}, overrides),
-        ])
+        txResult = await this.call<TransactionResponse>({
+          walletIndex,
+          contractAddress: assetAddress,
+          abi: erc20ABI,
+          funcName: 'transfer',
+          funcParams: [recipient, txAmount, Object.assign({}, overrides)],
+        })
       } else {
         // Transfer ETH
         const transactionRequest = Object.assign(
@@ -740,19 +743,14 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
    * @throws {"Failed to estimate gas price"} Thrown if failed to estimate gas price.
    */
   estimateGasPrices = async (): Promise<GasPrices> => {
-    let rates: FeeRates | undefined = undefined
-
     try {
-      rates = await this.getFeeRatesFromThorchain()
+      const rates = await this.getFeeRatesFromThorchain()
       return {
         average: baseAmount(rates.average, ETH_DECIMAL),
         fast: baseAmount(rates.fast, ETH_DECIMAL),
         fastest: baseAmount(rates.fastest, ETH_DECIMAL),
       }
-    } catch (error) {
-      console.log(error)
-      console.warn(`Error pulling rates from thorchain, will try alternate`)
-    }
+    } catch (error) {}
     //should only get here if thor fails
     return await this.estimateGasPricesFromEtherscan()
   }
@@ -787,7 +785,7 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
   /**
    * Estimate gas.
    *
-   * @param {FeesParams} params The transaction options.
+   * @param {FeesParams} params The transaction and fees options.
    * @returns {BaseAmount} The estimated gas fee.
    *
    * @throws {"Failed to estimate gas limit"} Thrown if failed to estimate gas limit.