+++ packages/xchain-ethereum/src/types/client-types.ts
@@ -42,9 +42,30 @@ export type FeesParams = C.FeesParams & C.TxParams
 export type FeesWithGasPricesAndLimits = { fees: C.Fees; gasPrices: GasPrices; gasLimit: BigNumber }
 
 export type ApproveParams = {
-  walletIndex: number
-  spender: Address
-  sender: Address
+  walletIndex?: number
+  contractAddress: Address
+  spenderAddress: Address
   feeOptionKey?: FeeOptionKey
   amount?: BaseAmount
+  // Optional fallback in case estimation for gas limit fails
+  gasLimitFallback?: ethers.BigNumberish
 }
+
+export type EstimateApproveParams = Omit<ApproveParams, 'feeOptionKey' | 'gasLimitFallback'>
+
+export type IsApprovedParams = {
+  walletIndex?: number
+  contractAddress: Address
+  spenderAddress: Address
+  amount?: BaseAmount
+}
+
+export type CallParams = {
+  walletIndex?: number
+  contractAddress: Address
+  abi: ethers.ContractInterface
+  funcName: string
+  funcParams?: Array<unknown>
+}
+
+export type EstimateCallParams = Pick<CallParams, 'contractAddress' | 'abi' | 'funcName' | 'funcParams' | 'walletIndex'>